requirements.txt
oldest-supported-numpy
cython==0.29.33

reno==3.3.0  # for changelog

setuptools>=46.4.0  # to support setup.cfg getting __version__
